+++ lib/cocoapods-core/podfile.rb
@@ -234,29 +234,39 @@ module Pod
 
     # Configures a new Podfile from the given ruby string.
     #
-    # @param  [String] string
-    #         The ruby string which will configure the podfile with the DSL.
-    #
     # @param  [Pathname] path
     #         The path from which the Podfile is loaded.
     #
+    # @param  [String] contents
+    #         The ruby string which will configure the podfile with the DSL.
+    #
     # @return [Podfile] the new Podfile
     #
-    def self.from_ruby(path)
-      string = File.open(path, 'r:utf-8') { |f| f.read }
+    def self.from_ruby(path, contents = nil)
+      contents ||= File.open(path, 'r:utf-8') { |f| f.read }
+
       # Work around for Rubinius incomplete encoding in 1.9 mode
-      if string.respond_to?(:encoding) && string.encoding.name != 'UTF-8'
-        string.encode!('UTF-8')
+      if contents.respond_to?(:encoding) && contents.encoding.name != 'UTF-8'
+        contents.encode!('UTF-8')
       end
+
+      if contents.tr!('“”‘’‛', %(""'''))
+        # Changes have been made
+        CoreUI.warn "Your #{path.basename} has had smart quotes sanitised. " \
+                    'To avoid issues in the future, you should not use ' \
+                    'TextEdit for editing it. If you are not using TextEdit, ' \
+                    'you should turn off smart quotes in your editor of choice.'
+      end
+
       podfile = Podfile.new(path) do
         # rubocop:disable Lint/RescueException
         begin
           # rubocop:disable Eval
-          eval(string, nil, path.to_s)
+          eval(contents, nil, path.to_s)
           # rubocop:enable Eval
         rescue Exception => e
           message = "Invalid `#{path.basename}` file: #{e.message}"
-          raise DSLError.new(message, path, e.backtrace, string)
+          raise DSLError.new(message, path, e.backtrace, contents)
         end
         # rubocop:enable Lint/RescueException
       end